If we're going for hot speed, let's not call this 7 times per checkin
I think the database probably won't change once this is in memory, so we
can call `is_postgres()` once and just store it.

If you're uncomfortable with it being an icky global, we could just do
the same thing in the `checkin` func and still cut down on some
redundency.