Update to new datatableview API, backpedal on URL changes
The datatables don't seem able to (without fooling around with them) to
retain the querystring values that are needed for filtering by
machine_group, business_unit, etc. This is due to generating the lengthy
querystring required for datatables and just replacing the request's
querystring.

Since the REST API is separate, and the views here need to only work
through the web app, AND we will never NOT want to filter by some group
(even 'all'), I restored the group_type and group_id values to the URL
paths. These are hardcoded and won't have any issues like filtering
inventory items by install path.

This also migrates to the new datatableview API, which splits up the
concept of a datatable from the actual view class that depends on it.