Update django to 2.0.1, along with others that need updating to load.
On my Mac at least, I had to add the `--no-binary psycopg2` option to
the requirements file for that package so that it would install.
Otherwise (my understanding) is that it conflicts with ssl.

http://initd.org/psycopg/docs/install.html#binary-install-from-pypi

To get pip freeze to work, I had to remove that line. I think it's
probably fine to leave out, but I wanted to make a note of it
somewhere.