Repopulate missing settings from default file when retrieved.
Slightly change `server.utils.get_setting` to update all settings that
have defaults in the `sal/default_sal_settings.json` file if a value is
missing when attempt to retrieve it. Only then fall back to any default
value given.

This prevents someone from deleting settings and having them just be
gone forever if the system is expecting them to be present for managing
behavior of the app.

This commit then also updates all invocations of `get_setting` to omit
the default argument where not needed.