Replace default User admin page with more useful one.
This commit removes the `UserProfile` admin page which was trying to
work around not knowing how to change the `User` admin page. After
figuring that out, it's redundent and less-featured.

The `is_superuser` property of User is suppressed because it doesn't
actually do anything useful in Sal at this time. That being said, some
documentation probably needs to be created to indicate what it is that
staff is doing (as in addition to admin page privileges, it also grants
user-editing permissions).

This also sorts some imports.