Rewrite plugin module heavily.
Rewrite Plugin subclasses to skip the inner Meta class.

This moves queryset instantiation and permissions checking to the plugin
rather than each view needing to handle it.

Create `OldPluginWrapper` which can be used with old plugins to make the
conform to the new Plugin API. This includes conversion of page names
('front' == 'all', etc).

Create a plugin manager to automatically wrap old-school plugins in a
wrapper that gives them the modern API, so other code doesn't have to
worry about it.