Add title property to Plugin, add `super_context` method. Refactor.
Title is by default constructed from the ClassName.

The `super_context` can be called by an overridden `get_context` method
to add in the calling kwargs, and the plugin, to the context. This just
saves some tedious copying of things like `group_id`, `group_type`, etc,
and should be useful for most all plugins (but necessary!)