Fix empty InstallReport, add report title block, restructure text_utils.
This started as just trying to fix InstallReport when there's no data so
it didn't look so broken.

Then I wanted to finish up the TODO item to add report titles and
descriptions in. To do so I ended up migrating all of the "text_utils"
out of utils into their own module to avoid a circular import.