Remove Plugin type and os families...
Plugin code had a `plugin_type` method and attribute which is no longer
needed since we can just use the Plugin's class itself for typing.

Plugin code continues to have an OS family attribute and method, but the
corresponding plugin/machine detail plugin/report database models that
track active/inactive state do not need to record this information, so
it was removed.