Revert sign; move version checking closer to consumer, only check 24h.
This removes checking for the newest Sal release during a machine
checkin.

There is no need to check for a new Sal version until a Sal GA user
visits the index page, because this data isn't used anywhere else. So I
removed it from checkin (and from the reporting checkin since it's not
used for either the client record or in @grahamgilbert's reports back).

I added the same mechanism for version checking to the
get_current_release_version_number thing, although by removing it from
the checkin function, it will be dramatically reduced to begin with. We
may just want to have it check more frequently or every time a GA user
visits index.

OR just move it to the maintenance script.