Remove out-of-date UpdateHistory and their items.
If an UpdateHistory is no longer a managed item (created every checkin
run from the Munki ManagedInstalls report), put it into a list for later
processing.

Then, for each of those, if it has no history items, delete it. Also, if
the most recent history item is older than our `HISTORICAL_DAYS`
setting, delete it.