Unroll InstalledUpdate, Pending[Apple]Update, and UpdateHistory process.
This code was too obtuse (I wrote it...)

So I unrolled it into separate sections so I could understand what was
going wrong.

What I discovered were a couple of problems:
1. PendingUpdate objects were not going to get created. I imagine
systems predating this code that stopped checking in would have some
left behind, but I don't see how they would have gotten created
otherwise. The problem was that it stopped once it determined it was
processing a ManagedInstall; but a ManagedInstall that is not installed
should generate a InstalledUpdate and a PendingUpdate record.
2. There were some incorrect assumptions about the format of the
different sections of the incoming data that made dictionary lookups
with the `get` method fall back to their default values.

After I started working on it, I realized that the PendingUpdate model
is not needed. The only clients of that model are InstallReport and
Pending3rdPartyUpdate plugins, both of which can be very easily updated
to just use InstalledUpdate filtered for `installed=False`.

So I rewrote the processing to no longer even consider creating
PendingUpdate records. I'll remove that model and all references in a
later commit.