Fix unique constraint on ManagedItem.name.
This need not, nor can be, unique, as every device will be creating
items for managed items that overlap. The uniqueness is handled by the
model's unique together setting.

This also removes the extra processing of Munki-submitted Apple Software
Update items that can result in constraint failures due to duplicate
items.