Rewrite PluginManager as a singleton / adapter and update clients.
This copies the relevent parts from Yapsy's impplementation, but due to
the use of name munging, it was cleaner to copy rather than subclass.

This commit then goes through all of the clients of the PluginManager
and updates them to do the right thing.

Since Sal does not use any of the info except for the plugin's path,
this commit removes the copying of attributes from the containing Yapsy
object aside from setting path, and it only does this once, to coincide
with only collecting plugins once (singleton!).