Slip the title into ajax view so we can `plugin.filter` just once.
This really cleans up the datatable template; pulled out all of the
stuff that's no longer really being used since datatables are handling
it.

Added the title (which was being just thrown away by the ajax call) into
the data it returns, and then used it in an initComplete func as part of
datatable setup to add the title in the existing place (outside of the
table).