Fix duplication of managed items from one source "over" another.
For the machine detail page, the ManagedItems display creates a dropdown
menu for sources that optionally provide multiple subtypes of items
(e.g. Munki with ManagedInstalls, ManagedUninstalls). Sources that don't
have these were having their items hidden due to the way the template
code was constructing intra-page anchor names. Specifically, the items
passed to the template used the key 'none' for the link href, and so if
you had more than one, they would get squashed.

This change adds the management source (slugified) as a prefix to the
item type to create unique targets for these links and solve this issue.