Increase searchDelay from implicit default of 400ms to 1000ms.
This controls how soon after a key release in the inventory search field
a query will be ajaxed to the server. 400ms is _pretty fast_, even for
fast typers, so we'll bump it up to 1000.

Of note, despite this, as soon as the first character is entered, it
seems like something starts happening, so this cooldown doesn't really
kick in until _after_ that.