Add js function to avoid query immediately after 1 char in inventory.
Despite the previous commit, datatables still does a query immediately
after a single character is entered in the search field. THEN it will do
the 1000ms pause before querying after subsequent input. This causes 2
queries (at least) when you should be able to get by with just one.

This code I yanked out of a very long SO thread about this. It will wait
to start searching until EITHER you have typed 3 characters, OR hit
Enter. AND it will also do a 1000ms pause.

This is nice because we also experimented with requiring an Enter key
hit to search, but it kills the experience that you would want from
using datatables in the first place. So now we can have both; if you
NEED to search for 1 or 2 characters, you can do so with Enter. And you
can also hit Enter whenever you feel like it to fire off the search.